Add more OOP
The bot now does everything via the Petal class and few other classes. Many process can
be abstracted under their own classes.

Create 4 new classes: Parser, Storage, TaskList, and Ui. Separate methods and fields into
respective classes.

Separating the bots into these classes allows me to model the bot as the interactions
between these components.
